+++ CHANGELOG.md
@@ -1,4 +1,3 @@
-## Version 4.0.1
+## Version 4.0.2
 
-- fixed a bug where casts of friendly players would show up when using Friendly Nameplates
-- fixed a bug where any alpha color change was not persisted
+- fixed a bug where Bars > Color: Target Class Color no longer worked